requirements.txt
python-dotenv
discord
horde-client
langchain